refactor: remover listagem /artigos (redundante com /busca)
A página /artigos era uma listagem filtrável quase idêntica à /busca —
mesma UI (ArticleFilters), mesmos filtros de data/órgão/tema, mesma
experiência. Com a landing como nova home e /noticias cobrindo a visão
curada, mantê-la só duplica a carga cognitiva dos usuários.

A página individual /artigos/[id] é preservada — continua sendo o destino
de cada card de notícia no portal inteiro (NewsCard, /noticias, /busca,
/orgaos, /temas, clippings, widget embed, feed RSS).

Mudanças:
- Deleta src/app/(public)/artigos/{page,ArticlesPageClient,actions}.tsx
- Move TagFacet type para src/types/article.ts (único consumidor real)
- Atualiza imports em ArticleFilters.tsx e TagFilter.tsx
- Redireciona links:
  * ClientArticle "Voltar às notícias": /artigos → /noticias
  * ClientArticle tags: /artigos?tags=X → /busca?q=X
  * ThemeTreeDisplay tema: /artigos?temas=X → /busca?temas=X
- e2e/zen-mode.authed.spec.ts: goto('/artigos') → goto('/busca')

Nota: TagFilter e a prop popularTags em ArticleFilters ficam como dead
code dormente (nenhum caller passa popularTags). Podem ser limpos depois.
