+++ src/components/articles/TagFilter.tsx
@@ -2,7 +2,7 @@
 
 import { ChevronDown, ChevronUp, Search, Tag, X } from 'lucide-react'
 import * as React from 'react'
-import type { TagFacet } from '@/app/(public)/artigos/actions'
+import type { TagFacet } from '@/types/article'
 
 type TagFilterProps = {
   popularTags: TagFacet[]
